Use main page layout with refined countdown
